+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Not yet released
 * Fix entity view pagination issue with joined collection and subselect fetching
 * Implement support for subselect fetching when main query builder has LIMIT/OFFSET and/or ORDER BY clause
 * Fix compatibility with Quarkus 2.1+ by registering `ValuesEntity` in deployment integration
+* Fix checking more uses of MULTISET fetching for proper type support
 
 ### Backwards-incompatible changes